requirements.txt
fastapi
httpx
uvicorn
python-dotenv
motor
python-jose[cryptography]
passlib[bcrypt]
